[RENAME/#14] otp -> pin 뷰 네이밍 변경
