+++ app/src/main/java/com/keyneez/presentation/login/LoginAdapter.kt
@@ -3,13 +3,13 @@ package com.keyneez.presentation.login
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.keyneez.presentation.login.otp.LoginOtpFragment
 import com.keyneez.presentation.login.phone.PhoneFragment
+import com.keyneez.presentation.login.pin.LoginPinFragment
 
 class LoginAdapter(fragmentActivity: FragmentActivity) : FragmentStateAdapter(fragmentActivity) {
     private val fragmentList = listOf(
         PhoneFragment(),
-        LoginOtpFragment()
+        LoginPinFragment()
     )
 
     override fun getItemCount(): Int = fragmentList.size
