+++ app/src/main/java/com/keyneez/presentation/login/pin/LoginPinFragment.kt
@@ -1,4 +1,4 @@
-package com.keyneez.presentation.login.otp
+package com.keyneez.presentation.login.pin
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,15 +7,15 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.viewModels
 import com.keyneez.presentation.login.LoginActivity
 import com.keyneez.presentation.main.MainActivity
-import com.keyneez.presentation.signup.otp.SignupOtpViewModel
+import com.keyneez.presentation.signup.pin.SignupPinViewModel
 import com.keyneez.util.binding.BindingFragment
 import com.keyneez.util.extension.hideKeyboard
 import com.keyneez.util.extension.setOnSingleClickListener
 import com.lab.keyneez.R
-import com.lab.keyneez.databinding.FragmentLoginOtpBinding
+import com.lab.keyneez.databinding.FragmentLoginPinBinding
 
-class LoginOtpFragment : BindingFragment<FragmentLoginOtpBinding>(R.layout.fragment_login_otp) {
-    private val viewModel by viewModels<SignupOtpViewModel>()
+class LoginPinFragment : BindingFragment<FragmentLoginPinBinding>(R.layout.fragment_login_pin) {
+    private val viewModel by viewModels<SignupPinViewModel>()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -27,11 +27,11 @@ class LoginOtpFragment : BindingFragment<FragmentLoginOtpBinding>(R.layout.fragm
     }
 
     private fun initHideKeyboard() {
-        binding.layoutLoginOtp.setOnClickListener { requireContext().hideKeyboard(requireView()) }
+        binding.layoutLoginPin.setOnClickListener { requireContext().hideKeyboard(requireView()) }
     }
 
     private fun initBackBtnClickListener() {
-        binding.btnLoginOtpBack.setOnSingleClickListener {
+        binding.btnLoginPinBack.setOnSingleClickListener {
             (activity as LoginActivity).intentToPreviousPage()
         }
     }
@@ -49,6 +49,6 @@ class LoginOtpFragment : BindingFragment<FragmentLoginOtpBinding>(R.layout.fragm
 
     companion object {
         @JvmStatic
-        fun newInstance() = LoginOtpFragment()
+        fun newInstance() = LoginPinFragment()
     }
 }
