+++ app/src/main/res/drawable/ic_signup_pin_erase.xml
