+++ app/src/main/res/values/strings.xml
@@ -9,9 +9,9 @@
     <string name="phone_number_hint">휴대폰 번호</string>
     <string name="phone_next">다음으로</string>
 
-    // login otp
-    <string name="login_otp_description">간편 비밀번호를\n입력해주세요.</string>
-    <string name="login_otp_rearrange">재배열</string>
+    // login pin
+    <string name="login_pin_description">간편 비밀번호를\n입력해주세요.</string>
+    <string name="login_pin_rearrange">재배열</string>
 
     // like
     <string name="like_save">저장</string>
@@ -112,12 +112,12 @@
     <string name="test_complete_start_keyneez">키니즈 시작하기</string>
     <string name="test_complete_my_item">나의 아이템</string>
 
-    // signup otp
-    <string name="signup_otp_description">간편 비밀번호를\n설정해주세요</string>
-    <string name="signup_otp_rearrange">재배열</string>
+    // signup pin
+    <string name="signup_pin_description">간편 비밀번호를\n설정해주세요</string>
+    <string name="signup_pin_rearrange">재배열</string>
 
-    // signup otp confirm
-    <string name="signup_otp_confirm_description">설정한 비밀번호를\n한번 더 입력해주세요</string>
+    // signup pin confirm
+    <string name="signup_pin_confirm_description">설정한 비밀번호를\n한번 더 입력해주세요</string>
 
 
     // onboarding
